+++ src/PlatformSettings/AppleFcmPlatformSettings.php
@@ -131,6 +131,12 @@ class AppleFcmPlatformSettings implements Arrayable
      */
     protected $launch_image;
 
+
+    /**
+     * @var ?bool
+     */
+    protected $mutable_content;
+
     /**
      * HTTP request headers defined in Apple Push Notification Service.
      * Refer to APNs request headers for supported headers, e.g. "apns-priority": "10".
@@ -293,6 +299,16 @@ class AppleFcmPlatformSettings implements Arrayable
         $this->launch_image = $launch_image;
     }
 
+    public function enableMutableContent(): void
+    {
+        $this->mutable_content = true;
+    }
+
+    public function disableMutableContent(): void
+    {
+        $this->mutable_content = null;
+    }
+
     /**
      * Build an array.
      *
@@ -318,6 +334,7 @@ class AppleFcmPlatformSettings implements Arrayable
                 'content-available' => $this->content_available,
                 'category' => $this->category,
                 'thread-id' => $this->thread_id,
+                'mutable-content' => $this->mutable_content,
             ],
         ];
     }